proxy/src/cli/pyproject.toml
[tool.poetry]
name = "synthlang"
version = "0.2.0"
description = "Command-line interface for the SynthLang framework using DSPy"
authors = ["SynthLang Team"]
readme = "README.md"
packages = [{include = "synthlang"}]

[tool.poetry.dependencies]
python = "^3.8"
dspy-ai = "^2.0.0"
click = "^8.1.7"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
rich = "^13.7.0"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
cryptography = "^41.0.5"
httpx = "^0.25.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
synthlang = "synthlang.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
